Uploading the initial Source Code
The two csproj project files were autogenerated by the ML.NET Model Builder (I mean they were autocreated and auto added to the solution, including source code). The only changes that i did in source code was in the main function (Program.cs) in order to receive one parameter (a json string) that contains the issue information that will used to predict. Also I added some comments in the same file in order to make the code easier to understand.
Also I added a .gitignore file in order to avoid uploading binaries and other files not needed in github repo.